Restructured the web-scrapers directory
