+++ web-scrapers/Dockerfile
@@ -8,4 +8,4 @@ RUN pip install -r requirements.txt
 
 COPY . .
 
-CMD ["python", "server.py"]
+CMD ["python", "src/server.py"]
