+++ web-scrapers/src/scraping_errors.py
