+++ web-scrapers/src/utils/requests_utils.py
@@ -3,6 +3,7 @@ import logging
 from scraping_errors import WebsiteScrapingError
 
 
+
 def check_whether_website_can_be_scraped(base_url: str) -> bool:
     status_code: int = requests.get(base_url).status_code
     if status_code == 200:
